Remove the unnecessary Question interface.
