+++ server/app/services/applicant/question/AddressQuestion.java
@@ -18,7 +18,7 @@ import services.question.types.QuestionType;
  *
  * <p>See {@link ApplicantQuestion} for details.
  */
-public class AddressQuestion extends QuestionImpl {
+public class AddressQuestion extends Question {
   private static final String PO_BOX_REGEX =
       "(?i)(.*(P(OST|.)?\\s*((O(FF(ICE)?)?)?.?\\s*(B(IN|OX|.?)))+)).*";
