+++ server/app/services/applicant/question/EnumeratorQuestion.java
@@ -15,7 +15,7 @@ import services.question.types.QuestionType;
  *
  * <p>See {@link ApplicantQuestion} for details.
  */
-public class EnumeratorQuestion extends QuestionImpl {
+public class EnumeratorQuestion extends Question {
 
   public EnumeratorQuestion(ApplicantQuestion applicantQuestion) {
     super(applicantQuestion);
