+++ server/app/services/applicant/question/FileUploadQuestion.java
@@ -15,7 +15,7 @@ import services.question.types.QuestionType;
  *
  * <p>See {@link ApplicantQuestion} for details.
  */
-public class FileUploadQuestion extends QuestionImpl {
+public class FileUploadQuestion extends Question {
 
   // This value is serving double duty as a singleton load of the value.
   // This value is an optional of an optional because not all questions are file upload questions,
