+++ server/app/services/applicant/question/SingleSelectQuestion.java
@@ -18,7 +18,7 @@ import services.question.types.QuestionType;
  *
  * <p>See {@link ApplicantQuestion} for details.
  */
-public class SingleSelectQuestion extends QuestionImpl {
+public class SingleSelectQuestion extends Question {
 
   // Stores the value, loading and caching it on first access.
   private Optional<Optional<LocalizedQuestionOption>> selectedOptionValueCache;
