+++ server/app/services/applicant/question/StaticContentQuestion.java
@@ -15,7 +15,7 @@ import services.question.types.QuestionType;
  *
  * <p>See {@link ApplicantQuestion} for details.
  */
-public class StaticContentQuestion extends QuestionImpl {
+public class StaticContentQuestion extends Question {
 
   public StaticContentQuestion(ApplicantQuestion applicantQuestion) {
     super(applicantQuestion);
